// frontend/projects/gideon/src/lib/api/model/applicationAction.ts
/**
 * Gideon API
 * Documentation Gideon API v1.0
 *
 * The version of the OpenAPI document: 1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface ApplicationAction {
  id: string;
  source?: string;
  target?: string;
  time: string;
  type: string;
  parameters?: object;
}